SWING-22 Update profile

+++ database/migrations/2024_07_30_194841_create_user_profiles_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->uuid('user_id');
             $table->foreign('user_id')->references('id')->on('users');
-            $table->integer('handicap')->nullable();
+            $table->decimal('handicap', 3, 1)->nullable();
             $table->timestamps();
         });
     }
